+++ examples/complete/variables.tf
@@ -3,3 +3,9 @@ variable "supporting_resources_name" {
   description = "Name of the supporting resources stack"
   default     = "terraform-aws-lambda"
 }
+
+variable "deletion_window_in_days" {
+  description = " The waiting period, specified in number of days. After the waiting period ends, AWS KMS deletes the KMS key. If you specify a value, it must be between 7 and 30, inclusive. If you do not specify a value, it defaults to 30. If the KMS key is a multi-Region primary key with replicas, the waiting period begins when the last of its replica keys is deleted. Otherwise, the waiting period begins immediately."
+  type        = number
+  default     = 7
+}