+++ CHANGELOG.md
@@ -9,11 +9,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: Add lambda event source mapping
 - feat: Add Lambda function URL
 - feat: Add Event Invoke Config
-- feat: Add lambda invocation feature
-- feat: Show more usage of the features in complete/specific feature examples
 - fix: CKV_AWS_50: "X-ray tracing is enabled for Lambda"
 - fix: CKV2_AWS_5: "Ensure that Security Groups are attached to another resource"
+- lambda example with ECR URI being the location containing the function's deployment package and add image configurations
+- add an example that has code_signing_config_arn
+- add an example with the following lambda permission attributes: qualifier, principle_org_id, statement_id_prefix, source_account, event_source_token, function_url_auth_type
 
+## [1.1.1] - 2023-10-26
+- fix: module's kms key
+- fix: security group ids
+- fix: lambda function timeouts block
+- lambda function example that uses external KMS key for env variables and CloudWatch encryption
+- added dead letter queue SQS to complete example
+- added file system configuration to complete example
+- removed alias routing configuration, the latest function version is already included in routing configuration.
+- refactored lambda layers and lambda layer permission blocks
+- lambda example with S3 bucket being the location containing the function's deployment package
+- modified lambda layer outputs
 
 ## [1.1.0] - 2023-09-04
 - feat: lambda invocation resource for the module
@@ -34,8 +46,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial commit
 - feat: add files from template repository
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-lambda/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-lambda/compare/1.1.1...HEAD
 
+[1.1.1]: https://github.com/boldlink/terraform-aws-lambda/releases/tag/1.1.1
 [1.1.0]: https://github.com/boldlink/terraform-aws-lambda/releases/tag/1.1.0
 [1.0.1]: https://github.com/boldlink/terraform-aws-lambda/releases/tag/1.0.1
 [1.0.0]: https://github.com/boldlink/terraform-aws-lambda/releases/tag/1.0.0